+++ src/sdk/vscode/CHANGELOG.md
@@ -12,6 +12,9 @@
 
 ## [Unreleased]
 
+- 0.9.5 [2021-06-07]
+  - Update certificate path within Kuksa.VAL repository
+
 - 0.9.3 [2021-03-09]
   - Add troubleshooting section in README.md
   - Fix version parsing for docker-compose
@@ -46,7 +49,7 @@
 
 - 0.7.0 [2021-02-12]
   - Add commands to facilitate the integration of Vehicle Signal Specification (VSS) into IoT Event Analytics
-    - Create complete Kuksa.VAL configuration
+    - Create complete Kuksa.val configuration
     - Convert VSS configuration to IoT Event Analytics types configuration
 
 - 0.6.1 [2021-02-10]